requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
backoff==2.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==14.0
coverage==7.6.0
croniter==2.0.5
dagster==1.7.15
dagster-graphql==1.7.15
dagster-pipes==1.7.15
dagster-webserver==1.7.15
dnspython==2.6.1
docstring_parser==0.16
filelock==3.15.4
fsspec==2024.6.1
gql==3.5.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.64.1
grpcio-health-checking==1.62.2
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
mypy==1.11.0
mypy-extensions==1.0.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pandas-stubs==2.2.2.240603
pendulum==3.0.0
pluggy==1.5.0
protobuf==4.25.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pymango==0.1.1
pymongo==4.8.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==70.3.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.38.2
structlog==24.2.0
tabulate==0.9.0
time-machine==2.14.2
tomli==2.0.1
toposort==1.10
tqdm==4.66.4
types-pytz==2024.1.0.20240417
typing_extensions==4.12.2
tzdata==2024.1
universal_pathlib==0.2.2
urllib3==2.2.2
uvicorn==0.30.4
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
